// src/app/models/ticket.ts
export class Ticket {
  name: string = '';
  surname: string = '';
  carrier: string = '';
  flightNo: string = '';
  classSeat: string = '';
  depAirport: string = '';
  arrAirport: string = '';
  dateOfBirth: string = '';
  luggage: string = '';
  seat: string = '';
  gate: string = '';
  boardingTime: string = '';
  passport: string = '';
}